requirements.txt
aws-cdk-lib==2.149.0
constructs>=10.0.0,<11.0.0
boto3>=1.17.0,<2.0.0

